+++ plugin_example/plugin_example/ConfigHandler.h
@@ -15,6 +15,6 @@ public:
 
 	__int64 getConfigProperty(std::string propertyName);
 
-	void setConfigProperty(std::string propertyName, int value);
+	void setConfigProperty(std::string propertyName, __int64 value);
 };
 #endif
